Change health endpoint from ping to health (#119)
Change health endpoint from ping to health and move health to status (#108)
