--- management_api_app/models/schemas/ping.py
@@ -1,7 +1,0 @@
-from datetime import datetime
-from pydantic import BaseModel
-
-
-class Pong(BaseModel):
-    message: str
-    time: datetime
