+++ management_api_app/services/health_checker.py
@@ -1,7 +1,7 @@
 from azure.core import exceptions
 from azure.cosmos import CosmosClient
 from core.config import STATE_STORE_ENDPOINT, STATE_STORE_KEY
-from models.schemas.health import StatusEnum
+from models.schemas.status import StatusEnum
 from resources import strings
